requirements.txt
streamlit>=1.32.0
openai>=1.12.0
pinecone
python-dotenv>=1.0.0
tiktoken>=0.5.2
PyPDF2>=3.0.0 